Fix clippy warning: casting integer literal to  is unnecessary

+++ aurora-standalone/refiner-types/src/bloom.rs
@@ -39,7 +39,7 @@ impl Bloom {
 
         for i in 0..BLOOM_BITS {
             let _ = i;
-            let mut index = 0 as usize;
+            let mut index = 0;
             for _ in 0..bloom_bytes {
                 index = (index << 8) | hash[ptr] as usize;
                 ptr += 1;
